refactor: organize templates by languages and simplify SourceType
