+++ devtools/common/src/main/java/io/quarkus/cli/commands/CreateProject.java
@@ -1,10 +1,5 @@
 package io.quarkus.cli.commands;
 
-import static io.quarkus.QuarkusTemplate.PROJECT_ARTIFACT_ID;
-import static io.quarkus.QuarkusTemplate.PROJECT_GROUP_ID;
-import static io.quarkus.QuarkusTemplate.PROJECT_VERSION;
-import static io.quarkus.QuarkusTemplate.QUARKUS_VERSION;
-import static io.quarkus.QuarkusTemplate.SOURCE_TYPE;
 import static io.quarkus.maven.utilities.MojoUtils.Element;
 import static io.quarkus.maven.utilities.MojoUtils.QUARKUS_VERSION_PROPERTY;
 import static io.quarkus.maven.utilities.MojoUtils.configuration;
@@ -13,6 +8,11 @@ import static io.quarkus.maven.utilities.MojoUtils.getPluginArtifactId;
 import static io.quarkus.maven.utilities.MojoUtils.getPluginGroupId;
 import static io.quarkus.maven.utilities.MojoUtils.getPluginVersion;
 import static io.quarkus.maven.utilities.MojoUtils.plugin;
+import static io.quarkus.templates.QuarkusTemplate.PROJECT_ARTIFACT_ID;
+import static io.quarkus.templates.QuarkusTemplate.PROJECT_GROUP_ID;
+import static io.quarkus.templates.QuarkusTemplate.PROJECT_VERSION;
+import static io.quarkus.templates.QuarkusTemplate.QUARKUS_VERSION;
+import static io.quarkus.templates.QuarkusTemplate.SOURCE_TYPE;
 
 import java.io.File;
 import java.io.IOException;
@@ -33,8 +33,8 @@ import org.apache.maven.model.PluginExecution;
 import org.apache.maven.model.PluginManagement;
 import org.apache.maven.model.Profile;
 
-import io.quarkus.SourceType;
 import io.quarkus.maven.utilities.MojoUtils;
+import io.quarkus.templates.SourceType;
 import io.quarkus.templates.TemplateRegistry;
 import io.quarkus.templates.rest.BasicRest;
