+++ devtools/common/src/main/java/io/quarkus/templates/SourceType.java
@@ -1,4 +1,4 @@
-package io.quarkus;
+package io.quarkus.templates;
 
 import io.quarkus.maven.utilities.MojoUtils;
