+++ devtools/common/src/main/resources/META-INF/services/io.quarkus.templates.QuarkusTemplate
